doc/design: Add design doc for object
The syscalls design phylosophy is to provide a unified interface for
the user to access the system resources, which is object handle. This
leads the object becomes a fundamental concept in the COCONUT-SVSM
kernel.

This document describes the object and object handle from both the user
mode and the COCONUT-SVSM kernel point of view.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>